+++ tests/appsec/appsec/test_telemetry.py
@@ -17,12 +17,13 @@ from ddtrace.internal.telemetry.constants import TELEMETRY_TYPE_GENERATE_METRICS
 from ddtrace.trace import tracer
 import tests.appsec.rules as rules
 from tests.appsec.utils import asm_context
+from tests.appsec.utils import build_payload
 from tests.utils import override_global_config
 
 
 config_asm = {"_asm_enabled": True}
 config_good_rules = {"_asm_static_rule_file": rules.RULES_GOOD_PATH, "_asm_enabled": True}
-invalid_rule_update = {"rules": {"test": "invalid"}}
+invalid_rule_update = [("ASM_DD", "Datadog/0/ASM/rules", {"rules": {"test": "invalid"}})]
 invalid_error = """appsec.waf.error::update::rules::bad cast, expected 'array', obtained 'map'"""
 
 
@@ -117,18 +118,15 @@ def test_metrics_when_appsec_block(telemetry_writer, tracer):
 def test_metrics_when_appsec_block_custom(telemetry_writer, tracer):
     telemetry_writer._namespace.flush()
     with asm_context(tracer=tracer, ip_addr=rules._IP.BLOCKED, span_name="test", config=config_asm) as span:
-        from ddtrace.appsec._remoteconfiguration import _appsec_rules_data
+        from ddtrace.appsec._remoteconfiguration import _appsec_callback
 
-        _appsec_rules_data(
-            {
-                "actions": [
-                    {
-                        "id": "block",
-                        "type": "block_request",
-                        "parameters": {"status_code": 429, "type": "json"},
-                    },
-                ]
-            },
+        actions = {
+            "actions": [{"id": "block", "type": "block_request", "parameters": {"status_code": 429, "type": "json"}}]
+        }
+        _appsec_callback(
+            [
+                build_payload("ASM", actions, "actions"),
+            ],
             tracer,
         )
         set_http_meta(span, rules.Config(), request_headers={"User-Agent": "Arachni/v1.5.1"})
@@ -200,11 +198,12 @@ unpatched_run = ddtrace.appsec._ddwaf.ddwaf_types.ddwaf_run
 
 
 def _wrapped_run(*args, **kwargs):
+    print(f"Wrapped run: {args} {kwargs}")
     unpatched_run(*args, **kwargs)
     return -3
 
 
-@mock.patch.object(ddtrace.appsec._ddwaf, "ddwaf_run", new=_wrapped_run)
+@mock.patch.object(ddtrace.appsec._ddwaf.waf, "ddwaf_run", new=_wrapped_run)
 def test_log_metric_error_ddwaf_internal_error(telemetry_writer):
     """Test that an internal error is logged when the WAF returns an internal error."""
     with override_global_config(dict(_asm_enabled=True, _asm_deduplication_enabled=False)):
